starpu: Add static cyclic graph version
This, like the explicit version, avoids MPI scalability concerns by only
going through the local part of the task graph.

It additionally avoids using the implicit dependency support from StarPU,
and instead submits to StarPU the task graph as a static cyclic task graph
using StarPU tags, which StarPU can then just run over and over, avoiding a
lot of implicit analysis overhead.

This depends on starpu_tag_notify_restart_from_apps, which is to be
introduced in StarPU 1.2.11, 1.3.5, and 1.4.0, and is available in their
respective branches in the meanwhile.